pyproject.toml
[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.build_meta"

[project]
name = "url2pdf-ai-reader"
version = "0.1.0"
description = "Convert URL content to AI-friendly PDF with FastAPI and Playwright."
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
authors = [{ name = "mobbb" }]
dependencies = [
  "fastapi==0.116.1",
  "uvicorn[standard]==0.35.0",
  "playwright==1.55.0",
  "beautifulsoup4==4.13.5",
  "readability-lxml==0.8.4.1",
  "lxml==6.0.2",
]

[project.urls]
Homepage = "https://github.com/REPLACE_ME/url2pdf-ai-reader"
Repository = "https://github.com/REPLACE_ME/url2pdf-ai-reader"

[project.scripts]
url2pdf-api = "app.cli:main"

[tool.setuptools]
packages = ["app"]
